requirements.txt
flask
flask-sqlalchemy
mysqlclient
flask-script
pymysql
request
uwsgi